Specify process environment using Python where possible
Change how the `Plex Media Scanner` is executed to allow specifying the
environment variables directly using an argument to the Popen call. This
removes the need to embed escaping the values from the configuration
file, allowing it to use the paths directly now.

A new configuration option (`PLEX_PYTHONHOME`) has been added allowing
you to specify the `PYTHONHOME` used when calling `Plex Media Scanner`.
This is necessary on some systems in order to properly run the scanner.

How `Plex Media Scanner` commands are built has been unified into a
command that simply takes the arguments to it, allowing the same logic
for how to run it to be specified in a single place instead of
duplicated in two locations.